This is the initial commit for the Galudon Character Calculator. This at current contains 3 pages, A character calculator page, a reference page, and a settings page. Each page has an header notifying the page, and a menu bar that follows the user down the page.
The character calculator page contains, A character information section, and 6 ability sections that contain their individual talents. When a talent is selected it updates numbers located on each ability section to display the proper calculations as per the system instructions.

The reference page takes that information and displays it in an easy to read and uneditable format that allows for quick reference to the character information, stats, and abilities.

The Settings page allows you to change certain guidelines related to the initial calculator.